fix(line3) renamed intersectToPlane -> intersectPointOfLineAndPlane
