+++ src/core/math/line3/intersectPointOfLineAndPlane.js
